feat(auth) : refactorisation de l'émission de jetons d'accès pour utiliser des objets DTO et suppression de la classe AccessTokenContext
