+++ api/src/RessAuth/Dto/AccessTokenPayload.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\RessAuth\Dto;
+
+final readonly class AccessTokenPayload
+{
+    public function __construct(
+        public string $accessToken,
+        public int $expiresIn,
+    ) {
+    }
+}
